Agrego mensajes de error cuando schedule task no esta presente

+++ src/service/schedule_task_service.py
@@ -29,6 +29,7 @@ def create(params, start_condition_type, restart_file_data, project_file_data, p
 
         return scheduled_task
 
+
 def update(_id, form):
     with get_session() as session:
         schedule_config = session.query(ScheduledTask).filter_by(id=_id).one_or_none()
@@ -42,7 +43,9 @@ def update(_id, form):
         schedule_config.geometry_id = form.geometry_id.data
         schedule_config.start_datetime = form.start_datetime.data
         if form.enabled.data:
-            if not file_storage_service.is_project_template_present(_id) or not file_storage_service.is_plan_template_present(_id) or not file_storage_service.is_restart_file_present(_id):
+            if not file_storage_service.is_project_template_present(
+                    _id) or not file_storage_service.is_plan_template_present(
+                _id) or not file_storage_service.is_restart_file_present(_id):
                 raise FileUploadError("Es mandatorio subir todos los archivos para habilitar la ejecución")
         schedule_config.enabled = form.enabled.data
         schedule_config.observation_days = form.observation_days.data
@@ -60,14 +63,16 @@ def update(_id, form):
         plan_file_service.process_plan_template(form.plan_file.data, _id)
         session.add(schedule_config)
 
+
 def _create_objects(old_objects, new_objects, create_func, _id=None):
     if not callable(create_func):
         raise ValueError("The create function provided is not callable")
-    
+
     if old_objects == []:
         return create_func(new_objects, _id)
 
-def _update_objects(old_objects, new_objects, update_func):    
+
+def _update_objects(old_objects, new_objects, update_func):
     if not callable(update_func):
         raise ValueError("The update function provided is not callable")
 
@@ -81,11 +86,12 @@ def _update_objects(old_objects, new_objects, update_func):
                 raise KeyError(error_msg) from ke
     return
 
+
 def update_objects(schedule_config, old_objects, new_objects, update_func, create_func, attr_name, _id=None):
     if old_objects == []:
         setattr(
-            schedule_config, 
-            attr_name, 
+            schedule_config,
+            attr_name,
             _create_objects(
                 old_objects,
                 new_objects,
@@ -95,51 +101,56 @@ def update_objects(schedule_config, old_objects, new_objects, update_func, creat
         )
     else:
         _update_objects(old_objects, new_objects, update_func)
-    
+
+
 def update_from_json(_id=None, **params):
     with get_session() as session:
-        schedule_config = session.query(ScheduledTask).filter_by(id = _id).one_or_none()
-        if schedule_config:
-            for key, value in params.items():
-                if value is not None:
-                    if key == 'border_conditions':
-                        update_objects(
-                            schedule_config,
-                            schedule_config.border_conditions,
-                            value,
-                            update_border_condition,
-                            process_series_json,
-                            'border_conditions',
-                            _id
-                        )
-                    elif key == 'plan_series_list':
-                        update_objects(
-                            schedule_config,
-                            schedule_config.plan_series_list,
-                            value,
-                            update_plan_series,
-                            process_plan_series_json,
-                            'plan_series_list',
-                            _id
-                        )
-                    elif key == 'initial_flows':
-                        update_objects(
-                            schedule_config,
-                            schedule_config.initial_flows, 
-                            value, 
-                            update_initial_flow, 
-                            process_initial_flows_json, 
-                            'initial_flows',
-                            _id
-                        )
-                    elif key == 'enabled':
-                        if value is True:
-                            if not file_storage_service.is_project_template_present(_id) or not file_storage_service.is_plan_template_present(_id) or not file_storage_service.is_restart_file_present(_id):
-                                raise FileUploadError("You must upload all the required files to enable execution")
-                        setattr(schedule_config, key, value)
-                    else:
-                        setattr(schedule_config, key, value)
+        schedule_config = session.query(ScheduledTask).filter_by(id=_id).one_or_none()
+        if not schedule_config:
+            raise Exception(f"Scheduled task with id {_id} not found")
+        for key, value in params.items():
+            if value is not None:
+                if key == 'border_conditions':
+                    update_objects(
+                        schedule_config,
+                        schedule_config.border_conditions,
+                        value,
+                        update_border_condition,
+                        process_series_json,
+                        'border_conditions',
+                        _id
+                    )
+                elif key == 'plan_series_list':
+                    update_objects(
+                        schedule_config,
+                        schedule_config.plan_series_list,
+                        value,
+                        update_plan_series,
+                        process_plan_series_json,
+                        'plan_series_list',
+                        _id
+                    )
+                elif key == 'initial_flows':
+                    update_objects(
+                        schedule_config,
+                        schedule_config.initial_flows,
+                        value,
+                        update_initial_flow,
+                        process_initial_flows_json,
+                        'initial_flows',
+                        _id
+                    )
+                elif key == 'enabled':
+                    if value is True:
+                        if not file_storage_service.is_project_template_present(
+                                _id) or not file_storage_service.is_plan_template_present(
+                            _id) or not file_storage_service.is_restart_file_present(_id):
+                            raise FileUploadError("You must upload all the required files to enable execution")
+                    setattr(schedule_config, key, value)
+                else:
+                    setattr(schedule_config, key, value)
         session.add(schedule_config)
+
 
 def update_files(scheduled_task_id, project_file=None, plan_file=None, restart_file=None):
     project_path = None
@@ -152,6 +163,7 @@ def update_files(scheduled_task_id, project_file=None, plan_file=None, restart_f
     if restart_file != None:
         restart_file_path = save_restart_file(restart_file, scheduled_task_id)
     return project_path, plan_path, restart_file_path
+
 
 def create_from_form(form, border_conditions, plan_series):
     params = {
@@ -188,9 +200,9 @@ def create_from_form(form, border_conditions, plan_series):
     return create(params, start_condition_type, restart_file_data, project_file_data, plan_file_data)
 
 
-
 def get_schedule_tasks(name=None, user_first_name=None, user_last_name=None, start_condition_type=None, date_from=None,
-                       date_to=None, enabled=None, frequency=None, calibration_id=None,calibration_id_for_simulations=None):
+                       date_to=None, enabled=None, frequency=None, calibration_id=None,
+                       calibration_id_for_simulations=None):
     with get_session() as session:
         query = session.query(ScheduledTask).order_by(ScheduledTask.id.desc())
         if name is not None:
@@ -217,6 +229,7 @@ def get_schedule_tasks(name=None, user_first_name=None, user_last_name=None, sta
             query = query.filter(ScheduledTask.calibration_id_for_simulations == calibration_id_for_simulations)
         return query.all()
 
+
 def get_schedule_task_config(schedule_config_id):
     with get_session() as session:
         return (
@@ -224,6 +237,7 @@ def get_schedule_task_config(schedule_config_id):
             .filter(ScheduledTask.id == schedule_config_id)
             .first()
         )
+
 
 def get_associated_execution_plans(scheduled_task_id):
     with get_session() as session:
@@ -233,16 +247,19 @@ def get_associated_execution_plans(scheduled_task_id):
             .all()
         )
 
+
 def delete_scheduled_task(scheduled_task_id):
     try:
         scheduled_task = get_schedule_task_config(scheduled_task_id)
+        if not scheduled_task:
+            raise Exception(f"Scheduled task with id {scheduled_task_id} not found")
         execution_plan_schedule_task_mapping = get_associated_execution_plans(scheduled_task_id)
 
         for mapping in execution_plan_schedule_task_mapping:
             execution_plan_service.delete_execution_plan(mapping.execution_id)
 
         file_storage_service.delete_scheduled_task(scheduled_task_id)
-            
+
         with get_session() as session:
             session.delete(scheduled_task)
             session.commit()
@@ -253,42 +270,44 @@ def delete_scheduled_task(scheduled_task_id):
         print(e)
         raise e
 
+
 def copy_schedule_task(scheduled_task_id):
     s = get_schedule_task_config(scheduled_task_id)
-    new_scheduled_task = create_copy(s.name, s.description, s.frequency, s.start_datetime, 
+    new_scheduled_task = create_copy(s.name, s.description, s.frequency, s.start_datetime,
                                      s.enabled, s.geometry_id, s.user_id, s.start_condition_type,
                                      s.observation_days, s.forecast_days, s.initial_flows,
                                      s.border_conditions, s.plan_series_list, s.calibration_id,
                                      s.calibration_id_for_simulations)
-    file_storage_service.copy_execution_files_scheduled(scheduled_task_id, new_scheduled_task.id) 
+    file_storage_service.copy_execution_files_scheduled(scheduled_task_id, new_scheduled_task.id)
     return new_scheduled_task
 
-def create_copy(name, description, frequency, start_datetime, 
+
+def create_copy(name, description, frequency, start_datetime,
                 enabled, geometry_id, user_id, start_condition_type,
                 observation_days, forecast_days, initial_flows,
                 border_conditions, plan_series_list, calibration_id,
                 calibration_id_for_simulations):
     with get_session() as session:
         scheduled_task = ScheduledTask(
-            name = name, 
-            description = description,
-            frequency = frequency,
-            start_datetime = start_datetime,
-            enabled = enabled,
-            geometry_id = geometry_id,
-            user_id = user_id,
-            start_condition_type = start_condition_type,
-            observation_days = observation_days,
-            forecast_days = forecast_days,
-            initial_flows = initial_flows,
-            border_conditions = border_conditions,
-            plan_series_list = plan_series_list,
-            calibration_id = calibration_id,
-            calibration_id_for_simulations = calibration_id_for_simulations
+            name=name,
+            description=description,
+            frequency=frequency,
+            start_datetime=start_datetime,
+            enabled=enabled,
+            geometry_id=geometry_id,
+            user_id=user_id,
+            start_condition_type=start_condition_type,
+            observation_days=observation_days,
+            forecast_days=forecast_days,
+            initial_flows=initial_flows,
+            border_conditions=border_conditions,
+            plan_series_list=plan_series_list,
+            calibration_id=calibration_id,
+            calibration_id_for_simulations=calibration_id_for_simulations
         )
-        
+
         session.add(scheduled_task)
         session.commit()
         session.refresh(scheduled_task)
-        
+
         return scheduled_task
